pyproject.toml
[tool.poetry]
name = "silero-tts"
version = "0.1.0"
description = "ahah"
authors = ["edpyt <edrsa1321s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"


[tool.poetry.group.http.dependencies]
fastapi = "^0.109.0"
uvicorn = "^0.27.0"


[tool.poetry.group.task.dependencies]
kink = "^0.7.0"


[tool.poetry.group.tts.dependencies]
silero = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
